+++ ui-components/index.ts
@@ -54,20 +54,22 @@ export * from "./src/headers/Heading"
 /* Helpers */
 export * from "./src/helpers/capitalize"
 export * from "./src/helpers/dateToString"
-export * from "./src/helpers/formOptions"
-export * from "./src/helpers/links"
-export * from "./src/helpers/mergeDeep"
-export * from "./src/helpers/numberOrdinal"
-export * from "./src/helpers/translator"
 export * from "./src/helpers/debounce"
-export * from "./src/helpers/validators"
+export * from "./src/helpers/formOptions"
 export * from "./src/helpers/formatYesNoLabel"
 export * from "./src/helpers/getTranslationWithArguments"
+export * from "./src/helpers/links"
+export * from "./src/helpers/mergeDeep"
+export * from "./src/helpers/MultiLineAddress"
+export * from "./src/helpers/numberOrdinal"
+export * from "./src/helpers/OneLineAddress"
 export * from "./src/helpers/preferences"
 export * from "./src/helpers/resolveObject"
+export * from "./src/helpers/tableSummaries"
+export * from "./src/helpers/translator"
 export * from "./src/helpers/useIntersect"
 export * from "./src/helpers/useMutate"
-export * from "./src/helpers/tableSummaries"
+export * from "./src/helpers/validators"
 
 /* Icons */
 export * from "./src/icons/HeaderBadge"
@@ -113,8 +115,6 @@ export * from "./src/page_components/listing/listing_sidebar/NumberedHeader"
 export * from "./src/page_components/listing/listing_sidebar/OrDivider"
 export * from "./src/page_components/listing/listing_sidebar/ReferralApplication"
 export * from "./src/page_components/listing/listing_sidebar/ContactAddress"
-export * from "./src/page_components/listing/listing_sidebar/OneLineAddress"
-export * from "./src/page_components/listing/listing_sidebar/MultiLineAddress"
 export * from "./src/page_components/listing/listing_sidebar/SubmitApplication"
 export * from "./src/page_components/listing/listing_sidebar/QuantityRowSection"
 export * from "./src/page_components/listing/listing_sidebar/WhatToExpect"
