+++ ui-components/src/global/text.scss
@@ -177,10 +177,3 @@ h1.title {
   @apply text-gray-700;
   @apply text-sm;
 }
-
-.sidebar-subheader {
-  @apply font-alt-sans;
-  @apply text-gray-750;
-  @apply text-sm;
-  @apply mb-3;
-}
