+++ ui-components/src/page_components/listing/listing_sidebar/ContactAddress.tsx
@@ -1,8 +1,8 @@
 import * as React from "react"
 import ReactDOMServer from "react-dom/server"
 import { Icon, IconFillColors } from "../../../icons/Icon"
-import { MultiLineAddress, Address } from "./MultiLineAddress"
-import { OneLineAddress } from "./OneLineAddress"
+import { MultiLineAddress, Address } from "../../../helpers/MultiLineAddress"
+import { OneLineAddress } from "../../../helpers/OneLineAddress"
 
 export interface ContactAddressProps {
   /** An address */
